+++ src/main/java/org/example/pensionatkademinaratingservice/dto/ReviewResponseDto.java
@@ -4,9 +4,9 @@ import java.time.LocalDate;
 
 public class ReviewResponseDto {
 
-    private Long customerId;
+    private int customerId;
     private String customerName; // logik för att hämta customer name.
-    private Long roomId;
+    private int roomId;
     private int rating;
     private String comment;
     private LocalDate date;
@@ -14,7 +14,7 @@ public class ReviewResponseDto {
     public ReviewResponseDto() {
     }
 
-    public ReviewResponseDto(Long customerId, String customerName, Long roomId, int rating, String comment, LocalDate date) {
+    public ReviewResponseDto(int customerId, String customerName, int roomId, int rating, String comment, LocalDate date) {
         this.customerId = customerId;
         this.customerName = customerName;
         this.roomId = roomId;
@@ -23,11 +23,11 @@ public class ReviewResponseDto {
         this.date = date;
     }
 
-    public Long getCustomerId() {
+    public int getCustomerId() {
         return customerId;
     }
 
-    public void setCustomerId(Long customerId) {
+    public void setCustomerId(int customerId) {
         this.customerId = customerId;
     }
 
@@ -39,11 +39,11 @@ public class ReviewResponseDto {
         this.customerName = customerName;
     }
 
-    public Long getRoomId() {
+    public int getRoomId() {
         return roomId;
     }
 
-    public void setRoomId(Long roomId) {
+    public void setRoomId(int roomId) {
         this.roomId = roomId;
     }
